Modularize sensors core
